Use separate hook for getting conversation ID
This change will prevent a bunch of unnecessary state updates and performance degradation.
