+++ wrappers/MessageInputWrapper.tsx
@@ -3,10 +3,12 @@ import { MessageInput } from "../component-library/components/MessageInput/Messa
 import { RecipientInputMode } from "../helpers";
 import useGetRecipientInputMode from "../hooks/useGetRecipientInputMode";
 import useSendMessage from "../hooks/useSendMessage";
+import useGetConversationId from "../hooks/useGetConversationId";
 
 export const MessageInputWrapper = () => {
   // XMTP Hooks
-  const { recipientInputMode, conversationId } = useGetRecipientInputMode();
+  const { recipientInputMode } = useGetRecipientInputMode();
+  const { conversationId } = useGetConversationId();
   const { sendMessage } = useSendMessage(conversationId as string);
 
   return (
